Added start of animation code
Updated the RenderComponent to handle animations. I will need to add a
StateComponent that likely will store an std::string to describe the
state and the RenderComponent will look at that if applicable to change
the animation. I will need to figure out where the frame update time
logic lives as well.