Adding Spaces for object and state management
First part of adding "Spaces" to the project which are conceptually
containers of systems and objects. Instead of global EntityFactorys and
ComponentManagers each Space will have their own. A space can be thought
of as a "layer" in the system and can be used for a lot of different
things such as different levels, menus vs gameplay and backgrounds vs
foregrounds etc. Anywhere we want to totally separate two sets of
entities. I have a lot more research to do on this but I like the
concept.