Ball Collisions pt 1
Starting to handle collisions with the ball. I realized that merely
setting dx and dy to values that will separate entities after the
MovementSystem gets to them doesn't quite work if we also don't clear
dx/dy. I.e. if the mandate is move 1 pixel right (dx = 1), the object
will move 1 pixel every second after having that collision because
nothing sets the dx. I'm thinking that VelocityComponents should have
separation and penetration to manually correct these values and to let
us to set dx/dy to 0 when appropriate (or other values if we want bounce
etc.)