Major Refactor pt.1
Fell off the horse a bit, but back in the game now. Doing a substantial
refactor to dip my toes into Data-Oriented Design. It's not a necessity
but I just wanted to get a sense of how to make data cache friendly.
Need to fix the collision detection though as I was still having issues
with that.